Follow Java Naming Conventions
Indentation should be 4 spaces not 2 spaces
Furthermore, removed unused variables and imports
